Refactoring [skip hazardlib][demos]
