+++ openquake/commonlib/readinput.py
@@ -425,48 +425,6 @@ def get_rupture(oqparam):
     return rup
 
 
-def read_source_groups(fname):
-    """
-    :param fname: a path to a source model XML file
-    :return: a list of SourceGroup objects containing source nodes
-    """
-    smodel = nrml.read(fname).sourceModel
-    src_groups = []
-    if smodel[0].tag.endswith('sourceGroup'):  # NRML 0.5 format
-        for sg_node in smodel:
-            sg = sourceconverter.SourceGroup(
-                sg_node['tectonicRegion'])
-            sg.sources = sg_node.nodes
-            src_groups.append(sg)
-    else:  # NRML 0.4 format: smodel is a list of source nodes
-        src_groups.extend(
-            sourceconverter.SourceGroup.collect(smodel))
-    return src_groups
-
-
-def get_source_ids(oqparam):
-    """
-    :param oqparam:
-        an :class:`openquake.commonlib.oqvalidation.OqParam` instance
-    :returns:
-        the complete set of source IDs found in all the source models
-    """
-    fnames = oqparam.inputs['source']
-    source_ids = set()
-    logging.info('Reading source IDs from %d model file(s)', len(fnames))
-    for fname in fnames:
-        if fname.endswith('.hdf5'):
-            with hdf5.File(fname, 'r') as f:
-                for sg in f['/']:
-                    for src in sg:
-                        source_ids.add(src.source_id)
-        else:
-            for sg in read_source_groups(fname):
-                for src_node in sg:
-                    source_ids.add(src_node['id'])
-    return source_ids
-
-
 def get_source_model_lt(oqparam):
     """
     :param oqparam:
@@ -480,8 +438,7 @@ def get_source_model_lt(oqparam):
         # NB: converting the random_seed into an integer is needed on Windows
         return logictree.SourceModelLogicTree(
             fname, validate=False, seed=int(oqparam.random_seed),
-            num_samples=oqparam.number_of_logic_tree_samples,
-            source_ids=get_source_ids(oqparam))
+            num_samples=oqparam.number_of_logic_tree_samples)
     return logictree.FakeSmlt(oqparam.inputs['source_model'],
                               int(oqparam.random_seed),
                               oqparam.number_of_logic_tree_samples)
@@ -538,7 +495,7 @@ def get_source_models(oqparam, gsim_lt, source_model_lt, monitor,
                 src_groups.extend(psr.parse(
                     fname, pik[fname], apply_unc, oqparam.investigation_time))
             else:  # just collect the TRT models
-                src_groups.extend(read_source_groups(fname))
+                src_groups.extend(logictree.read_source_groups(fname))
         num_sources = sum(len(sg.sources) for sg in src_groups)
         sm.src_groups = src_groups
         trts = [mod.trt for mod in src_groups]
